additional documentation to accommodate a blendshape workflow
dna_viewer_grab_changes_from_scene_and_propagate_to_dna.py

A lot of devs and artists want to use a custom metahuman face in place of the original metahuman. This script allows that but the documentation is needs a little more elaboration. 
There is also a minor bug fix that can trip up a lot of devs. 
This will hopefully make things a little more clear until Epic release an updated version of this script.